requirements.txt
Flask
openai>=1.0.0
PyMuPDF
beautifulsoup4
requests
matplotlib
python-dotenv
